refactor: correct more types
