+++ src/servers/Cargo.toml
@@ -55,7 +55,7 @@ once_cell = "1.16"
 openmetrics-parser = "0.4"
 opensrv-mysql = "0.4"
 parking_lot = "0.12"
-pgwire = { git = "https://github.com/sunng87/pgwire", branch = "feature/async-query-parser" }
+pgwire = "0.15"
 pin-project = "1.0"
 postgres-types = { version = "0.2", features = ["with-chrono-0_4"] }
 promql-parser = "0.1.1"
